Making Changes for WriteList and PKUpdateOperationHandler.java
The PKUpdateOperation Handler was updated so that if a non primary key record/file is updated with the same information two consecutive times the record isn't deleted in MarkLogic.

WriteList:
This updated fixes an issue with updating primary key records. With the addition of the writting each document with the word root to the file/record on updates it was causing the document not to be able to compary the Map Key value pairs since the comparison was taking place at the incorrect level.